Quantitative strategy and its backtest
Using the Combined Decision Trees to predict whether the stock would go limit up or limit down in the next day. If limit up, mark it as 1, otherwise 0, which are the buy in and sell out signal respectively. To Train our predictor, we need high frequency order book data of all the stock in the market.